+++ tests/Unit/DSC_PfxImport.Tests.ps1
@@ -44,6 +44,8 @@ try
             -TypeName System.Management.Automation.PSCredential `
             -ArgumentList $testUsername, (ConvertTo-SecureString $testPassword -AsPlainText -Force)
 
+        $certificateContent = [System.Convert]::ToBase64String(@(00, 00, 00))
+
         $validPath = "TestDrive:\$testFile"
         $validCertPath = "Cert:\LocalMachine\My"
         $validCertFullPath = '{0}\{1}' -f $validCertPath, $validThumbprint
@@ -84,7 +86,14 @@ try
                 $Store -eq 'My'
         }
 
-        $importPfxCertificate_parameterfilter = {
+        $importPfxCertificateWithContent_parameterfilter = {
+            $CertStoreLocation -eq $validCertPath -and `
+                $Base64Content -eq $certificateContent -and `
+                $Exportable -eq $True -and `
+                $Password -eq $testCredential.Password
+        }
+
+        $importPfxCertificateWithFile_parameterfilter = {
             $CertStoreLocation -eq $validCertPath -and `
                 $FilePath -eq $validPath -and `
                 $Exportable -eq $True -and `
@@ -115,6 +124,17 @@ try
             Verbose    = $True
         }
 
+        $presentParamsWithContent = @{
+            Thumbprint = $validThumbprint
+            Content    = $certificateContent
+            Ensure     = 'Present'
+            Location   = 'LocalMachine'
+            Store      = 'My'
+            Exportable = $True
+            Credential = $testCredential
+            Verbose    = $True
+        }
+
         $presentParamsWithFriendlyName = @{
             Thumbprint   = $validThumbprint
             Path         = $validPath
@@ -127,7 +147,73 @@ try
             FriendlyName = $certificateFriendlyName
         }
 
+        $presentParamsWithFriendlyNameWithContent = @{
+            Thumbprint   = $validThumbprint
+            Content      = $certificateContent
+            Ensure       = 'Present'
+            Location     = 'LocalMachine'
+            Store        = 'My'
+            Exportable   = $True
+            Credential   = $testCredential
+            Verbose      = $True
+            FriendlyName = $certificateFriendlyName
+        }
+
+        $presentParamsWithoutContentAndPath = @{
+            Thumbprint   = $validThumbprint
+            Ensure       = 'Present'
+            Location     = 'LocalMachine'
+            Store        = 'My'
+            Exportable   = $True
+            Credential   = $testCredential
+            Verbose      = $True
+            FriendlyName = $certificateFriendlyName
+        }
+
+        $presentParamsWithBothContentAndPath = @{
+            Thumbprint   = $validThumbprint
+            Content      = $certificateContent
+            Path         = $validPath
+            Ensure       = 'Present'
+            Location     = 'LocalMachine'
+            Store        = 'My'
+            Exportable   = $True
+            Credential   = $testCredential
+            Verbose      = $True
+            FriendlyName = $certificateFriendlyName
+        }
+
+        $absentParamsWithBothContentAndPath = @{
+            Thumbprint   = $validThumbprint
+            Content      = $certificateContent
+            Path         = $validPath
+            Ensure       = 'Absent'
+            Location     = 'LocalMachine'
+            Store        = 'My'
+            Exportable   = $True
+            Credential   = $testCredential
+            Verbose      = $True
+            FriendlyName = $certificateFriendlyName
+        }
+
         $absentParams = @{
+            Thumbprint = $validThumbprint
+            Ensure     = 'Absent'
+            Location   = 'LocalMachine'
+            Store      = 'My'
+            Verbose    = $True
+        }
+
+        $absentParamsWithContent = @{
+            Thumbprint = $validThumbprint
+            Content    = $certificateContent
+            Ensure     = 'Absent'
+            Location   = 'LocalMachine'
+            Store      = 'My'
+            Verbose    = $True
+        }
+
+        $absentParamsWithoutContentAndPath = @{
             Thumbprint = $validThumbprint
             Ensure     = 'Absent'
             Location   = 'LocalMachine'
@@ -199,7 +285,7 @@ try
 
                 It 'Should not throw exception' {
                     {
-                        $script:result = Get-TargetResource @presentParams
+                        $script:result = Get-TargetResource @presentParamsWithContent
                     } | Should -Not -Throw
                 }
 
@@ -209,7 +295,8 @@ try
 
                 It 'Should contain the input values' {
                     $script:result.Thumbprint | Should -BeExactly $validThumbprint
-                    $script:result.Path | Should -BeExactly $validPath
+                    $script:result.Path | Should -BeExactly ''
+                    $script:result.Content | Should -Be $presentParamsWithContent.Content
                     $script:result.Ensure | Should -BeExactly 'Absent'
                     $script:result.FriendlyName | Should -BeNullOrEmpty
                 }
@@ -224,6 +311,34 @@ try
         }
 
         Describe 'DSC_PfxImport\Test-TargetResource' -Tag 'Test' {
+            Context 'When Content and Path parameters are null' {
+                It 'Should throw exception when Ensure is Present' {
+                    {
+                        Test-TargetResource @presentParamsWithoutContentAndPath
+                    } | Should -Throw
+                }
+
+                It 'Should not throw exception when Ensure is Absent' {
+                    {
+                        Test-TargetResource @absentParamsWithoutContentAndPath
+                    } | Should -Not -Throw
+                }
+            }
+
+            Context 'When both Content and Path parameters are set' {
+                It 'Should throw exception when Ensure is Present' {
+                    {
+                        Test-TargetResource @presentParamsWithBothContentAndPath
+                    } | Should -Throw ($script:localizedData.ContentAndPathParametersAreSet)
+                }
+
+                It 'Should not throw exception when Ensure is Absent' {
+                    {
+                        Test-TargetResource @absentParamsWithBothContentAndPath
+                    } | Should -Not -Throw
+                }
+            }
+
             Context 'When certificate is not in store but should be' {
                 Mock -CommandName Get-CertificateFromCertificateStore
 
@@ -281,8 +396,37 @@ try
             BeforeAll {
                 Mock -CommandName Test-Path -MockWith { $true }
                 Mock -CommandName Import-PfxCertificate
+                Mock -CommandName Import-PfxCertificateEx
                 Mock -CommandName Remove-CertificateFromCertificateStore
                 Mock -CommandName Set-CertificateFriendlyNameInCertificateStore
+            }
+
+            Context 'When Content and Path parameters are null' {
+                It 'Should throw exception when Ensure is Present' {
+                    {
+                        Set-TargetResource @presentParamsWithoutContentAndPath
+                    } | Should -Throw ($script:localizedData.ContentAndPathParametersAreNull)
+                }
+
+                It 'Should not throw exception when Ensure is Absent' {
+                    {
+                        Set-TargetResource @absentParamsWithoutContentAndPath
+                    } | Should -Not -Throw
+                }
+            }
+
+            Context 'When both Content and Path parameters are set' {
+                It 'Should throw exception when Ensure is Present' {
+                    {
+                        Set-TargetResource @presentParamsWithBothContentAndPath
+                    } | Should -Throw ($script:localizedData.ContentAndPathParametersAreSet)
+                }
+
+                It 'Should not throw exception when Ensure is Absent' {
+                    {
+                        Set-TargetResource @absentParamsWithBothContentAndPath
+                    } | Should -Not -Throw
+                }
             }
 
             Context 'When PFX file exists and certificate should be in the store but is not' {
@@ -304,8 +448,82 @@ try
                 It 'Should call Import-PfxCertificate with expected parameters' {
                     Assert-MockCalled `
                         -CommandName Import-PfxCertificate `
-                        -ParameterFilter $importPfxCertificate_parameterfilter `
+                        -ParameterFilter $importPfxCertificateWithFile_parameterfilter `
                         -Exactly -Times 1
+                }
+
+                It 'Should not call Import-PfxCertificateEx' {
+                    Assert-MockCalled -CommandName Import-PfxCertificateEx -Exactly -Times 0
+                }
+
+                It 'Should not call Set-CertificateFriendlyNameInCertificateStore' {
+                    Assert-MockCalled -CommandName Set-CertificateFriendlyNameInCertificateStore -Exactly -Times 0
+                }
+
+                It 'Should not call Remove-CertificateFromCertificateStore' {
+                    Assert-MockCalled -CommandName Remove-CertificateFromCertificateStore -Exactly -Times 0
+                }
+            }
+
+            Context 'When PFX file exists and certificate should be in the store but is not (No Import-PfxCertificate cmdlet)' {
+                Mock -CommandName Get-CertificateFromCertificateStore
+                Mock -CommandName Test-CommandExists -MockWith { $false }
+
+                It 'Should not throw exception' {
+                    {
+                        Set-TargetResource @presentParams
+                    } | Should -Not -Throw
+                }
+
+                It 'Should call Test-Path with expected parameters' {
+                    Assert-MockCalled `
+                        -CommandName Test-Path `
+                        -ParameterFilter $testPath_parameterfilter `
+                        -Exactly -Times 1
+                }
+
+                It 'Should call Import-PfxCertificateEx with expected parameters' {
+                    Assert-MockCalled `
+                        -CommandName Import-PfxCertificateEx `
+                        -ParameterFilter $importPfxCertificateWithFile_parameterfilter `
+                        -Exactly -Times 1
+                }
+
+                It 'Should not call Import-PfxCertificate' {
+                    Assert-MockCalled -CommandName Import-PfxCertificate -Exactly -Times 0
+                }
+
+                It 'Should not call Set-CertificateFriendlyNameInCertificateStore' {
+                    Assert-MockCalled -CommandName Set-CertificateFriendlyNameInCertificateStore -Exactly -Times 0
+                }
+
+                It 'Should not call Remove-CertificateFromCertificateStore' {
+                    Assert-MockCalled -CommandName Remove-CertificateFromCertificateStore -Exactly -Times 0
+                }
+            }
+
+            Context 'When PFX content is used and certificate should be in the store but is not' {
+                Mock -CommandName Get-CertificateFromCertificateStore
+
+                It 'Should not throw exception' {
+                    {
+                        Set-TargetResource @presentParamsWithContent
+                    } | Should -Not -Throw
+                }
+
+                It 'Should call Test-Path' {
+                    Assert-MockCalled -CommandName Test-Path -Exactly -Times 0
+                }
+
+                It 'Should call Import-PfxCertificate with expected parameters' {
+                    Assert-MockCalled `
+                        -CommandName Import-PfxCertificateEx `
+                        -ParameterFilter $importPfxCertificateWithContent_parameterfilter `
+                        -Exactly -Times 1
+                }
+
+                It 'Should not call Import-PfxCertificate' {
+                    Assert-MockCalled -CommandName Import-PfxCertificate -Exactly -Times 0
                 }
 
                 It 'Should not call Set-CertificateFriendlyNameInCertificateStore' {
@@ -327,11 +545,42 @@ try
                     } | Should -Not -Throw
                 }
 
-                It 'Should not call Test-Path with expected parameters' {
+                It 'Should not call Test-Path' {
                     Assert-MockCalled -CommandName Test-Path -Exactly -Times 0
                 }
 
-                It 'Should not call Import-PfxCertificate with expected parameters' {
+                It 'Should not call Import-PfxCertificate' {
+                    Assert-MockCalled -CommandName Import-PfxCertificate -Exactly -Times 0
+                }
+
+                It 'Should not call Set-CertificateFriendlyNameInCertificateStore' {
+                    Assert-MockCalled -CommandName Set-CertificateFriendlyNameInCertificateStore -Exactly -Times 0
+                }
+
+                It 'Should not call Remove-CertificateFromCertificateStore' {
+                    Assert-MockCalled -CommandName Remove-CertificateFromCertificateStore -Exactly -Times 0
+                }
+            }
+
+            Context 'When PFX content is used and certificate should be in the store and is' {
+                Mock -CommandName Get-CertificateFromCertificateStore `
+                    -MockWith $validCertificateWithPrivateKey_mock
+
+                It 'Should not throw exception' {
+                    {
+                        Set-TargetResource @presentParamsWithContent
+                    } | Should -Not -Throw
+                }
+
+                It 'Should not call Test-Path' {
+                    Assert-MockCalled -CommandName Test-Path -Exactly -Times 0
+                }
+
+                It 'Should not call Import-PfxCertificateEx' {
+                    Assert-MockCalled -CommandName Import-PfxCertificateEx -Exactly -Times 0
+                }
+
+                It 'Should not call Import-PfxCertificate' {
                     Assert-MockCalled -CommandName Import-PfxCertificate -Exactly -Times 0
                 }
 
@@ -374,6 +623,40 @@ try
                 }
             }
 
+            Context 'When PFX content is used and certificate should be in the store and is and the friendly name is different' {
+                Mock -CommandName Get-CertificateFromCertificateStore `
+                    -MockWith $validCertificateWithDifferentFriendlyName_mock
+
+                It 'Should not throw exception' {
+                    {
+                        Set-TargetResource @presentParamsWithFriendlyNameWithContent
+                    } | Should -Not -Throw
+                }
+
+                It 'Should not call Test-Path' {
+                    Assert-MockCalled -CommandName Test-Path -Exactly -Times 0
+                }
+
+                It 'Should not call Import-PfxCertificateEx' {
+                    Assert-MockCalled -CommandName Import-PfxCertificateEx -Exactly -Times 0
+                }
+
+                It 'Should not call Import-PfxCertificate' {
+                    Assert-MockCalled -CommandName Import-PfxCertificate -Exactly -Times 0
+                }
+
+                It 'Should call Set-CertificateFriendlyNameInCertificateStore with expected parameters' {
+                    Assert-MockCalled `
+                        -CommandName Set-CertificateFriendlyNameInCertificateStore `
+                        -ParameterFilter $setCertificateFriendlyNameInCertificateStore_parameterfilter `
+                        -Exactly -Times 1
+                }
+
+                It 'Should not call Remove-CertificateFromCertificateStore' {
+                    Assert-MockCalled -CommandName Remove-CertificateFromCertificateStore -Exactly -Times 0
+                }
+            }
+
             Context 'When PFX file exists and certificate should be in the store and is and the friendly name is the same' {
                 Mock -CommandName Get-CertificateFromCertificateStore `
                     -MockWith $validCertificateWithPrivateKey_mock
@@ -384,11 +667,15 @@ try
                     } | Should -Not -Throw
                 }
 
-                It 'Should not call Test-Path with the parameters supplied' {
+                It 'Should not call Test-Path' {
                     Assert-MockCalled -CommandName Test-Path -Exactly -Times 0
                 }
 
-                It 'Should not call Import-PfxCertificate with the parameters supplied' {
+                It 'Should not call Import-PfxCertificateEx' {
+                    Assert-MockCalled -CommandName Import-PfxCertificateEx -Exactly -Times 0
+                }
+
+                It 'Should not call Import-PfxCertificate' {
                     Assert-MockCalled -CommandName Import-PfxCertificate -Exactly -Times 0
                 }
 
@@ -401,6 +688,36 @@ try
                 }
             }
 
+            Context 'When PFX content is used and certificate should be in the store and is and the friendly name is the same' {
+                Mock -CommandName Get-CertificateFromCertificateStore `
+                    -MockWith $validCertificateWithPrivateKey_mock
+
+                It 'Should not throw exception' {
+                    {
+                        Set-TargetResource @presentParamsWithFriendlyNameWithContent
+                    } | Should -Not -Throw
+                }
+
+                It 'Should not call Test-Path' {
+                    Assert-MockCalled -CommandName Test-Path -Exactly -Times 0
+                }
+
+                It 'Should not call Import-PfxCertificateEx' {
+                    Assert-MockCalled -CommandName Import-PfxCertificateEx -Exactly -Times 0
+                }
+
+                It 'Should not call Import-PfxCertificate' {
+                    Assert-MockCalled -CommandName Import-PfxCertificate -Exactly -Times 0
+                }
+
+                It 'Should not call Set-CertificateFriendlyNameInCertificateStore' {
+                    Assert-MockCalled -CommandName Set-CertificateFriendlyNameInCertificateStore -Exactly -Times 0
+                }
+
+                It 'Should not call Remove-CertificateFromCertificateStore' {
+                    Assert-MockCalled -CommandName Remove-CertificateFromCertificateStore -Exactly -Times 0
+                }
+            }
 
             Context 'When PFX file exists and certificate should not be in the store but is' {
                 Mock -CommandName Get-CertificateFromCertificateStore `
@@ -414,6 +731,44 @@ try
 
                 It 'Should not call Test-Path' {
                     Assert-MockCalled -CommandName Test-Path -Exactly -Times 0
+                }
+
+                It 'Should not call Import-PfxCertificateEx' {
+                    Assert-MockCalled -CommandName Import-PfxCertificateEx -Exactly -Times 0
+                }
+
+                It 'Should not call Import-PfxCertificate' {
+                    Assert-MockCalled -CommandName Import-PfxCertificate -Exactly -Times 0
+                }
+
+                It 'Should not call Set-CertificateFriendlyNameInCertificateStore' {
+                    Assert-MockCalled -CommandName Set-CertificateFriendlyNameInCertificateStore -Exactly -Times 0
+                }
+
+                It 'Should call Remove-CertificateFromCertificateStore with expected parameters' {
+                    Assert-MockCalled `
+                        -CommandName Remove-CertificateFromCertificateStore `
+                        -ParameterFilter $removeCertificateFromCertificateStore_parameterfilter `
+                        -Exactly -Times 1
+                }
+            }
+
+            Context 'When PFX content is used and certificate should not be in the store but is' {
+                Mock -CommandName Get-CertificateFromCertificateStore `
+                    -MockWith $validCertificateWithPrivateKey_mock
+
+                It 'Should not throw exception' {
+                    {
+                        Set-TargetResource @absentParamsWithContent
+                    } | Should -Not -Throw
+                }
+
+                It 'Should not call Test-Path' {
+                    Assert-MockCalled -CommandName Test-Path -Exactly -Times 0
+                }
+
+                It 'Should not call Import-PfxCertificateEx' {
+                    Assert-MockCalled -CommandName Import-PfxCertificateEx -Exactly -Times 0
                 }
 
                 It 'Should not call Import-PfxCertificate' {
@@ -443,6 +798,43 @@ try
 
                 It 'Should not call Test-Path' {
                     Assert-MockCalled -CommandName Test-Path -Exactly -Times 0
+                }
+
+                It 'Should not call Import-PfxCertificateEx' {
+                    Assert-MockCalled -CommandName Import-PfxCertificateEx -Exactly -Times 0
+                }
+
+                It 'Should not call Import-PfxCertificate' {
+                    Assert-MockCalled -CommandName Import-PfxCertificate -Exactly -Times 0
+                }
+
+                It 'Should not call Set-CertificateFriendlyNameInCertificateStore' {
+                    Assert-MockCalled -CommandName Set-CertificateFriendlyNameInCertificateStore -Exactly -Times 0
+                }
+
+                It 'Should call Remove-CertificateFromCertificateStore' {
+                    Assert-MockCalled `
+                        -CommandName Remove-CertificateFromCertificateStore `
+                        -ParameterFilter $removeCertificateFromCertificateStore_parameterfilter `
+                        -Exactly -Times 1
+                }
+            }
+
+            Context 'When PFX content is used and certificate should not be in the store and is not' {
+                Mock -CommandName Get-CertificateFromCertificateStore
+
+                It 'Should not throw exception' {
+                    {
+                        Set-TargetResource @absentParamsWithContent
+                    } | Should -Not -Throw
+                }
+
+                It 'Should not call Test-Path' {
+                    Assert-MockCalled -CommandName Test-Path -Exactly -Times 0
+                }
+
+                It 'Should not call Import-PfxCertificateEx' {
+                    Assert-MockCalled -CommandName Import-PfxCertificateEx -Exactly -Times 0
                 }
 
                 It 'Should not call Import-PfxCertificate' {
@@ -476,6 +868,43 @@ try
                     Assert-MockCalled `
                         -CommandName Test-Path `
                         -ParameterFilter $testPath_parameterfilter `
+                        -Exactly -Times 1
+                }
+
+                It 'Should not call Import-PfxCertificateEx' {
+                    Assert-MockCalled -CommandName Import-PfxCertificateEx -Exactly -Times 0
+                }
+
+                It 'Should not call Import-PfxCertificate' {
+                    Assert-MockCalled -CommandName Import-PfxCertificate -Exactly -Times 0
+                }
+
+                It 'Should not call Set-CertificateFriendlyNameInCertificateStore' {
+                    Assert-MockCalled -CommandName Set-CertificateFriendlyNameInCertificateStore -Exactly -Times 0
+                }
+
+                It 'Should not call Remove-CertificateFromCertificateStore' {
+                    Assert-MockCalled -CommandName Remove-CertificateFromCertificateStore -Exactly -Times 0
+                }
+            }
+
+            Context 'When PFX content is not valid and certificate should be in the store' {
+                Mock -CommandName Import-PfxCertificateEx -MockWith { throw }
+
+                It 'Should throw exception' {
+                    {
+                        Set-TargetResource @presentParamsWithContent
+                    } | Should -Throw
+                }
+
+                It 'Should not call Test-Path' {
+                    Assert-MockCalled -CommandName Test-Path -Exactly -Times 0
+                }
+
+                It 'Should call Import-PfxCertificateEx with expected parameters' {
+                    Assert-MockCalled `
+                        -CommandName Import-PfxCertificateEx `
+                        -ParameterFilter $importPfxCertificateWithContent_parameterfilter `
                         -Exactly -Times 1
                 }